Fix transaction creation methods signatures with double type

+++ BankAccount/BankAccount/Compte.cs
@@ -25,13 +25,13 @@ namespace BankAccount.Test
             return _historiqueDeTransactions.BalanceActuelle();
         }
 
-        public void Deposer(int montant, DateTime date)
+        public void Deposer(double montant, DateTime date)
         {
             var nouveauDepot = TransactionBancaire.CreerDepot(montant, date, BalanceActuelle());
             _historiqueDeTransactions.AjouterTransaction(nouveauDepot);
         }
 
-        public void Retirer(int montant, in DateTime date)
+        public void Retirer(double montant, in DateTime date)
         {
             var nouveauRetrait = TransactionBancaire.CreerRetrait(montant, date, BalanceActuelle());
             _historiqueDeTransactions.AjouterTransaction(nouveauRetrait);
